+++ src/logic/systems/collision.hpp
@@ -6,20 +6,20 @@ public:
     Collision(int x, int y, int width, int height);
     ~Collision();
 
-    int getX() const { return x; }
-    int getY() const { return y; }
+    float getX() const { return x; }
+    float getY() const { return y; }
 
     int getWidth() const { return width; }
     int getHeight() const { return height; }
 
     void setXYPosition(int x, int y);
-    void setXPosition(int x);
-    void setYPosition(int y);
+    void setXPosition(float x);
+    void setYPosition(float y);
 
 private:
     float x;
     float y;
-    float width;
-    float height;
+    int width;
+    int height;
 
 };
